# kbot-final/pkg/sdk/openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "kbot 企业级 AI Agent 教学平台 REST API，由 swaggo/swag 从 handler 注解生成。",
        "title": "kbot API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/skills": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "skills"
                ],
                "summary": "列出技能",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "skills"
                ],
                "summary": "从 SKILL.md 创建技能",
                "parameters": [
                    {
                        "description": "SKILL.md",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/skill.CreateSkillRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/skills/{skill_id}/versions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "skills"
                ],
                "summary": "列出 Skill 不可变版本",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Skill ID",
                        "name": "skill_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "skills"
                ],
                "summary": "创建 Skill 不可变版本",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Skill ID",
                        "name": "skill_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "SKILL.md",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.CreateSkillVersionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "skill.CreateSkillRequest": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "skill_md": {
                    "description": "完整 SKILL.md 文本",
                    "type": "string"
                },
                "workspace_id": {
                    "type": "string"
                }
            }
        },
...
        "v1.CreateSkillVersionRequest": {
            "type": "object",
            "properties": {
                "skill_md": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}